sentinal_ml/requirements.txt
fastapi==0.135.3
kafka_python==2.3.0
mlflow==2.10.2
pandas<3
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.33.1
scikit_learn==1.8.0
xgboost==3.2.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-kafka-python==0.45b0